Feature/arpa 344

+++ src/app/features/dashboards/user-list/user-list.component.ts
@@ -1,16 +1,16 @@
 import { Observable } from 'rxjs';
 import { TranslateService } from '@ngx-translate/core';
-import { Component, EventEmitter, Input, Output, OnChanges, SimpleChanges } from '@angular/core';
+import { Component, EventEmitter, Input, OnChanges, Output, SimpleChanges } from '@angular/core';
 import { orderBy } from 'lodash-es';
 import { ConfirmationService, TreeNode } from 'primeng/api';
 import { OverlayPanel } from 'primeng/overlaypanel';
-import { IUserDto } from '../../../models/IUserDto';
-import { IRoleDto } from '../../../models/IRoleDto';
-import { ISectionTreeDto } from '../../../models/section';
-import { ISetRoleDto } from '../../../models/ISetRoleDto';
 import { UserService } from '../../../core/services/user.service';
 import { NotificationsService } from '../../../core/services/notifications.service';
 import { AuthService } from '../../../core/services/auth.service';
+import { UserDto } from '../../../model/userDto';
+import { SectionTreeDto } from '../../../model/sectionTreeDto';
+import { SetRoleDto } from '../../../model/setRoleDto';
+import { RoleDto } from '../../../model/roleDto';
 
 @Component({
   selector: 'arpa-user-list',
@@ -18,13 +18,13 @@ import { AuthService } from '../../../core/services/auth.service';
   styleUrls: ['./user-list.component.scss'],
 })
 export class UserListComponent implements OnChanges {
-  @Input() users: IUserDto[] | null = [];
-  @Input() roles: IRoleDto[] = [];
-  @Input() sectionTree: ISectionTreeDto | undefined;
+  @Input() users: UserDto[] | null = [];
+  @Input() roles: RoleDto[] = [];
+  @Input() sectionTree: SectionTreeDto | undefined;
   @Output() userDeleted = new EventEmitter<string>();
-  @Output() rolesSet = new EventEmitter<ISetRoleDto>();
-  usersWithoutRole: IUserDto[] | undefined = [];
-  selectedUser: IUserDto | null = null;
+  @Output() rolesSet = new EventEmitter<string>();
+  usersWithoutRole: UserDto[] | undefined = [];
+  selectedUser: UserDto | null = null;
   selectedRoles: string[] = [];
   maxRoleLevel$: Observable<number>;
   selectedUserSections: TreeNode[] = [];
@@ -45,24 +45,28 @@ export class UserListComponent implements OnChanges {
     }
   }
 
-  getInitials(user: IUserDto): string {
-    return `${user.displayName
-      .split(' ')
-      .map((name) => name.length && name[0].toUpperCase() || '')
-      .join('')}`;
+  getInitials(user: UserDto): string {
+    if (user.displayName) {
+      return `${user.displayName
+        .split(' ')
+        .map((name) => name.length && name[0].toUpperCase() || '')
+        .join('')}`;
+    } else {
+      return '';
+    }
   }
 
   onSelected(event: any, panel: any): void {
     this.selectedUser = event.option;
-    this.selectedRoles = this.selectedUser!.roleNames;
+    this.selectedRoles = this.selectedUser!.roleNames || [];
     this.selectedUserSections = this.mapSectionTree(this.sectionTree!, event.option.stakeholderGroupIds);
     panel.toggle(event.originalEvent);
   }
 
   saveRoles(panel: OverlayPanel): void {
-    const dto: ISetRoleDto = { userName: this.selectedUser!.userName, roleNames: this.selectedRoles };
+    const dto: SetRoleDto = { userName: this.selectedUser!.userName, roleNames: this.selectedRoles };
     this.authService.setUserRoles(dto).subscribe(() => {
-      this.rolesSet.emit(dto);
+      this.rolesSet.emit(dto.userName);
       panel.hide();
       this.notificationsService.success('userlist.USER_ROLES_SET');
     });
@@ -88,25 +92,20 @@ export class UserListComponent implements OnChanges {
   getSections(stakeholderGoupIds: string[]): TreeNode[] {
     return this.mapSectionTree(this.sectionTree!, stakeholderGoupIds);
   }
-
-  private reset(): void {
-    this.selectedUser = null;
-    this.selectedRoles = [];
-    this.selectedUserSections = [];
-  }
-
   private deleteSelectedUser(): void {
-    this.userService.deleteUser(this.selectedUser!.userName).subscribe(() => {
-      this.userDeleted.emit(this.selectedUser!.userName);
-      this.notificationsService.success('userlist.USER_DELETED');
-    });
+    if (this.selectedUser!.userName) {
+      this.userService.deleteUser(this.selectedUser!.userName).subscribe(() => {
+        this.userDeleted.emit(this.selectedUser!.userName);
+        this.notificationsService.success('userlist.USER_DELETED');
+      });
+    }
   }
 
-  private mapSectionTree(sectionTree: ISectionTreeDto, stakeholderGoupIds: string[]): TreeNode[] {
+  private mapSectionTree(sectionTree: SectionTreeDto, stakeholderGoupIds: string[]): TreeNode[] {
     return sectionTree.children.map((child) => this.mapSectionNode(child, stakeholderGoupIds));
   }
 
-  private mapSectionNode(sectionNode: ISectionTreeDto, stakeholderGoupIds: string[]): TreeNode {
+  private mapSectionNode(sectionNode: SectionTreeDto, stakeholderGoupIds: string[]): TreeNode {
     return {
       data: sectionNode.data,
       children: sectionNode.children.map((child) => this.mapSectionNode(child, stakeholderGoupIds)),
